+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Transports that don't have unambiguous targets are defined with `selection` as `explicit` so they won't be automatically used unless selected explicitly on a notification, or configuration overridden
   - Deliveries no longer have a `default_` prefix, although existing automations which use these will automatically be switched to `email`,`notify_entity` etc
   - SMS auto discovery works for Twilio and Mikrotik SMS integrations
+  - A repair is raised for any deliveries that have the same name as a transport but don't use that transport. The transport name is effectively a reserved delivery name.
 - Creating `Delivery` objects now only necessary if there's more than one Delivery for the same transport, like `plain_email` and `html_email`, different Telegram channels etc
     - Everything that can be done with a `Delivery` configuration can be done with the `delivery_defaults:` section of a `Transport` object
     - Its also possible to avoid creating Delivery objects in YAML by defining the relevant `data:` items in a notification action, although this gets unwieldy (the whole point of Delivery objects is to define this stuff once and not across many notifications).